Revert "Add more unit tests.  better nan and empty column handling"
